proportion ok,manque les distances egales et fish ball

+++ srcs/isitwall/ft_closer.c
@@ -6,7 +6,7 @@
 /*   By: pganglof <pganglof@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/29 17:04:39 by pganglof          #+#    #+#             */
-/*   Updated: 2019/12/05 15:39:51 by pganglof         ###   ########.fr       */
+/*   Updated: 2019/12/05 18:26:32 by pganglof         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@ t_pos	*ft_closer(t_map *map, t_pos *wall_hor, t_pos *wall_ver)
 			printf("I wall_hor->posx : %d\nI wall_hor->posy : %d\n", wall_hor->posx, wall_hor->posy);
 		//printf("map->mapx : %d\n", map->x);
 		//printf("map->mapy : %d\n", map->y);
-		wall_ver->dis = sqrt(pow(map->gamer->unitx - wall_ver->unitx, 2)
-		+ pow(map->gamer->unity - wall_ver->unity, 2));
+		wall_ver->dis = sqrt(pow((double)map->gamer->unitx - (double)wall_ver->unitx, 2.0)
+		+ pow((double)map->gamer->unity - (double)wall_ver->unity, 2.0));
 		return (free_wall(wall_ver, &wall_hor));
 	}
 	if (wall_ver->posx < 0 || wall_ver->posx >= map->x
@@ -34,14 +34,14 @@ t_pos	*ft_closer(t_map *map, t_pos *wall_hor, t_pos *wall_ver)
 			printf("II wall_ver->posx : %d\nII wall_ver->posy : %d\n", wall_ver->posx, wall_ver->posy);
 		//printf("map->mapx : %d\n", map->x);
 		//printf("map->mapy : %d\n", map->y);
-		wall_hor->dis = sqrt(pow(map->gamer->unitx - wall_hor->unitx, 2)
-		+ pow(map->gamer->unity - wall_hor->unity, 2));
+		wall_hor->dis = sqrt(pow((double)map->gamer->unitx - (double)wall_hor->unitx, 2.0)
+		+ pow((double)map->gamer->unity - (double)wall_hor->unity, 2.0));
 		return (free_wall(wall_hor, &wall_ver));
 	}
-	wall_hor->dis = pow(map->gamer->unitx - wall_hor->unitx, 2)
-	+ pow(map->gamer->unity - wall_hor->unity, 2);
-	wall_ver->dis = pow(map->gamer->unitx - wall_ver->unitx, 2)
-	+ pow(map->gamer->unity - wall_ver->unity, 2);
+	wall_hor->dis = pow((double)map->gamer->unitx - (double)wall_hor->unitx, 2.0)
+	+ pow((double)map->gamer->unity - (double)wall_hor->unity, 2.0);
+	wall_ver->dis = pow((double)map->gamer->unitx - (double)wall_ver->unitx, 2.0)
+	+ pow((double)map->gamer->unity - (double)wall_ver->unity, 2.0);
 	if (map->gamer->beta == 30)
 	{
 	if (wall_hor->dis == wall_ver->dis)
